Cleanups and fixes
git-svn-id: https://svn.ntop.org/svn/ntop/trunk/nDPIng@8158 a02cd0c1-8e76-42e1-a119-56f9641475e2

+++ src/ndpi/lib/scanners/proto_ayiya.c
@@ -28,7 +28,7 @@
 */
 
 
-#include "ndpi_protocols.h"
+#include "ndpi_api.h"
 
 struct ayiya {
   u_int8_t flags[3];
@@ -54,7 +54,7 @@ void ndpi_search_ayiya(struct ndpi_detection_module_struct *ndpi_struct, struct 
       u_int32_t fireyears = 86400 * 365 * 5;
     
       #ifndef __KERNEL__
-	now = time(NULL);
+	now = (u_int32_t)time(NULL);
       #else
 	now = 1402729042; /* Dummy workaround */
       #endif
